ts3-query
deliverData function will be split into deliverClient and deliverServer,
which splits the respective files regarding client and server
information into blocks. if no user is online, there is no client data,
which leads to errors, so this split will fix that issue

due to testing events the ${clientDataDir} was not removed in the
cleanup function after the script ran. the folder will be removed again