[ i3 global config ] cookie-clicker click script
if you push $mod + RightShift + LeftShift you will start the click
script. you probably have to customize it to your needs

if you push $mod + LeftShift + RightShift you'll *kill all watch
instances*, so if you're watching ${something} it will be killed as
well! keep that in mind, if you want to use that