[ screen rotation ] remapping key
due to malfunctions with my display buttons to rotate the screen, i will
use the launch1 button instead.

the `rotateScreen.zsh` script will be invoked twice if the screen is
`normal`, but only once, if it is inverted.

the script works expected when invoked from the shell or via any other
button, but not via one of my two buttons on the display (lenovo x230t).