[ mutt + isync ] completely revamping the config file(s)
up to this point, each mailbox was manually added to the mbsyncrc,
causing a lot of hand work and was prone to typing errors.

likewise, each of those mailboxes was added via the `mailboxes` command
into the account specific muttrc.

from now on, only the accounts are ... "querried" by mbsync, i.e. *all*
of the imap folders will be downloaded.

to fill mutt's sidebars again, there is a script
`mutt-sidebar-fill.bash`, called each time a account specific muttrc is
sourced, which effectively means: start mutt or change the account will
trigger the script to run on the set $folder. the output is a string
containing each folder (except `/new`, `/tmp` and `/cur`) and some
regex-foo making them add the folder as a mailbox.

e.g. if the script finds a folder called `Inbox/Drafs`, it will be
reformatted to `mailboxes =Inbox/Drafs`.

beside that, all smtp connections will be switched to "smtps", which
seems like a good idea to me :)

yeah, i know, this is a (way too) large commit, but it's my personal
dotfiles repository, so who will blame me ;-)

the mailcap will change as well. `zathura` will be replaced by `mupdf`
and for svg graphics, `eog` will take point

the sidebar will now look a bit different, mainly because i switched the
$sidebar_short_path option on to make the sidebar's readability better.

the indent string will be edited to 3 instead of 2 whitespaces as well
and some unnecessary color options will be deleted, due to their
presence in the color scheme.

since i switched to neomutt, the progressbar will be colored :-) i just
need proper colors, so technically speaking, these configs aren't
compatible to mutt anymore.