[ mutt | create config ] removing password foo
before this time, i was juggling with password files, gpg and stuff to
put one plaintext file with one password for each account i have into a
folder on my system, where the account-specific muttrc files sourced the
correct password file. due to using gpg now, these files are not
necessary anymore.

this script should now run on every system, if it doesn't: well ... :o)

the script will not test anymore, whether some specific programmes are
installed. if the paths/folders are not accessible, the programmes I
used in this script will probably call for attention.

the script will exit, though, if the password file is not accessible.

- removing the gpg decrypt/encrypt stuff
- the copyFiles function will not ask for any input anymore, since every
time the whole lot of files are tried to copy into their places
- the script will make sure, the `~/.mutt/passwords` folder exists
- it is not of importance, if the script was issued before
- the script will not check for system's architecture anymore (`uname -s`)
- the script will not run mbsync automatically anymore